CenterNet/requirements.txt
torch
torchvision
tensorboard
scipy==1.2.1
numpy==1.17.0
matplotlib==3.1.2
# opencv_python==4.1.2.30
opencv_python
tqdm==4.60.0
Pillow==8.2.0
h5py==2.10.0

ExifRead
ipywidgets==7.6.5
lmdb
# mmcv
# opencv_python_headless==4.5.5.64
opencv_python_headless
pandas
plotly==5.15.0
pybind11==2.11.1
pytest==6.2.4
pytorch_msssim==0.2.1
PyYAML==6.0
range_coder==1.1
rawpy==0.18.1
seaborn==0.11.2
setuptools==58.0.4
thop==0.1.1.post2207130030
tifffile==2021.7.2
torchac
scikit_image

